backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cryptography==42.0.7
pyotp==2.9.0
lxml==5.2.1
aiofiles==23.2.1
python-dotenv==1.0.1
httpx==0.27.0
babel==2.14.0
